+++ src/crawlbrulee/types/__init__.py
@@ -17,6 +17,7 @@ from .async_ import (
 from .common import (
     ApiErrorName,
     ApiErrorResponse,
+    BillingEngine,
     ProxyTier,
     RateLimitErrorDetails,
     ResolvedProxyTier,
@@ -71,6 +72,7 @@ __all__ = [
     # common
     "ProxyTier",
     "ResolvedProxyTier",
+    "BillingEngine",
     "ScreenshotType",
     "ScreenshotDeviceMode",
     "ScreenshotCleanup",